Seperate MenuController
