+++ resources/views/admin/layouts/master.blade.php
@@ -127,7 +127,7 @@
             };
 
             $.ajaxSetup({ headers: { 'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content') } });
-            $.ajax({ dataType: 'json', timeout: 3000, url: "{!! route('admin.getmenu') !!}" + "?id=admin" })
+            $.ajax({ dataType: 'json', timeout: 3000, url: "{!! route('getmenu') !!}" + "?id=admin" })
             .done ( function(data, textStatus, jqXHR) { 
                 const $top = $("#topMenu");
                 $.each( data.menu, function ( index, top ) {
